Improve UX for loading interactive chunk popovers.
An interactive chunk's mouseover popover could be stuck open if the mouse moved over a region that would trigger the popover, but moved off before the data used to populate that popover was loaded from the server. This would cause the popover to be shown after the mouseout event had already fired, resulting in the popover being stuck open until the next time the mouse was hovered over the region that spawned the popover in the first place.

The fix is to switch the popover from data-trigger="hover" to data-trigger="manual", then manage whether the popover should be shown using our own mouseover and mouseout events and keeping track of whether a popover _should_ be shown at a given time.