Reload contents of interactive chunk popover when shown.
Each popover's content was previously loaded from the server only once, so if a popover was previously instantiated for a chunk, then the server-side data changed, the popover for that chunk would display outdated data.

The new behaviour is to load new content from the server each time, showing the previously-cached content as a placeholder until new content has been loaded.

The initial load UX has also been improved, as the user now sees a "Loading..." placeholder instead of simply no popover at all if the server is taking a while to respond.