Implement ability to prompt user for input
Prompting the user for input will replace the terminal's standard prompt with some custom text. The content entered by the user in response to a prompt will be treated as a response to that prompt, and is not processed as a command and not added to input history.

The user's input in response to the prompt can optionally be masked. For example, each character typed into a password may be masked as "*".

For example, to prompt a user for login credentials:

this.promptInput("Username:")
	.then((response: string) => {
		username = response;
		return this.promptInput("Password:", "*");
	})
	.then((response: string) => {
		password = response;
		return this.sessionService.login(username, password);
	})
	.then(() => {
		// Handle successful login
	});

In support of the above functionality, implemented a MaskPipe for masking strings in templates.