Improve UX for logout and login
Logging out now unsubscribes from the event stream, and informs the user to type "login" to log back in.

The local "login" command now prompts for a username and password instead of requiring the user to type their username and password as part of the command line.

Removed some obsolete debugging console.log()s.