Implement EmbedMediaComponent
Currently, the purpose of this component is to act as a replacement for the <img> tag. If a wob is associated with an image, the binary data for that image is retrievable as a wob property, but wob properties can't be accessed without an authorization token. This class fetches the requested image by providing the authorization header, then converts the image into a data: URL to be embedded onto the page.

Wob hover popovers now show image properties.

In the future, EmbedMediaComponent may be extended to suport other kinds of embeddable media, both from the Flowerbox server and from third-party services.