Scroll to bottom when interactive chunk layout changed.
The terminal scrolls to the bottom after each chunk is inserted. However, if an interactive chunk has content that is loaded asynchronously (such as an image), then the height of the chunk may be different after the content has loaded. When this happens, the terminal needs to be re-scrolled to the bottom.