Autocomplete fixes.
The word list generator was infinitely looping if the start of the command line is a space.

The matching function was failing to prepend the command line if the last-known matches were found at the end of the command line after a space. That caused the command-thus-far to be deleted when matches were returned, since the longest string in common between all matches was "".