Implement sudo support.
Preceding a terminal command with "sudo" will cause the command to be executed in a privileged context. This works by obtaining an admin-privileged authorization token from the server, and providing the admin token in the API call Authorization header instead of the regular token.

SessionService is now able to store a token and adminToken separately, as well as obtain (login) and invalidate (logout) each token separately.

When making an API call via SessionHttp, the "options" argument may now include an "admin" property. If options.admin is truthy, then the API call is submitted using SessionService.adminToken instead of SessionService.token. On token error, only the token that was provided with the request that failed is invalidated.