WobService: Cleaner way of getting binary properties.
An X-Property-Metadata header is now available from the server when requesting a binary property. The header contains a JSON string with the metadata that would have been included in the body of the response when requesting a non-binary property. Using this metadata, it is no longer necessary to rely on the getInfo() response to fill in parts of a Property object.

getBinaryProperty() now returns a Property object with the Blob content of the response encoded as a data URL in the "value" property.