Refactor property editing.
GenericPropertyEditorComponent is now responsible for contacting the server to save and delete drafts and properties.

PropertyEditorComponent is responsible for maintaining a list of known properties, as the list is required for showing a GenericPropertyEditorComponent-derived component for each property.

Also fixed a bug where a draft with a boolean value would be saved as an integer value instead when applied.