GenericPropertyEditor: Use separate member variable as input to PermissionsComponent.
Angular 2's OnChanges lifecycle hook only triggers when an input property's value has changed. When the value is an reference to an object, then the lifecycle hook only triggers if the reference is pointed at a different object.

In order for PermissionsComponent to reflect changes in permission, it must be aware when the Property or Verb object whose permissions it is displaying has changed. But it can only be aware of the change if the reference to the Property or Verb object has changed.

Modifying GenericPropertyEditor.property to point at a new instance of the Property object worked for this purpose, but it also caused the reference of the Property object being passed into GenericPropertyEditor itself to change (via PropertyEditor.onChange(), in response to GenericPropertyEditor.propertyChange). Since GenericPropertyEditor is instantiated by PropertyEditor as part of an ngFor loop, a change in the reference to the input Property causes the corresponding instance of GenericPropertyEditor to be destroyed and replaced. That was in turn causing changes to the property draft to not be saved, because GenericPropertyEditor is destroyed before the debounce timeout that would trigger a call to saveDraft() expires.

To cirvumvent all that, we now pass a separate object reference, GenericPropertyEditor._propertyForPermissions, to PermissionsComponent as an input property. GenericPropertyEditor._propertyForPermissions gets replaced with a new instance whenever GenericPropertyEditor.property is modified, and GenericPropertyEditor.property itself retains a reference to the same Property object with which it was initialised.