Verb constructor: A zero-length sigs array is valid input.
It means a draft exists for the signatures component, and that it contains zero signatures. This may happen, for example, if the applied version of the verb has one or more signatures, and all signatures were deleted by the editor, but the draft storing those deletions hasn't yet been committed to the applied version.